Docker_laddernet/requirements.txt
# Name                    Version
pytorch                   1.13.1
torchvision               0.14.1
cudatoolkit               11.8.0
matplotlib                3.5.3
numpy                     1.21.6
opencv                    4.6.0
pandas                    1.3.5
pillow                    9.2.0
scikit-learn              1.0.2
scipy                     1.7.3
tensorboard               2.11.2
tqdm                      4.66.5
threadpoolctl             3.1.0
h5py
config